Require Jenkins 2.346.3 or newer
https://issues.jenkins.io/browse/JENKINS-70286 notes that the job copy
builder plugin has an implied dependency on the WMI Windows Agents plugin
due to the age of the Jenkins version on which it dependes.

https://plugins.jenkins.io/jobcopy-builder/#dependencies shows that it
has many other implied dependencies due to that old Jenkins version.

This change resolves almost all of those implied dependencies (except
instance identity, it was split in Jenkins 2.357).

https://www.jenkins.io/doc/developer/plugin-development/dependency-management/#jenkins-plugin-bom
describes the plugin bill of materials that is used in this pull request.
It simplifies dependency version management.